# orders/src/main/resources/static/swagger.yaml
openapi: 3.0.1
info:
  version: "1.0.0"
  title: Orders API
  description: API para la gestión de órdenes
servers:
  - url: http://localhost:8080
    description: Local server
paths:
  /api/orders:
    get:
      summary: Obtener órdenes
      description: Retorna una lista de órdenes
      responses:
        '200':
          description: Lista de órdenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      summary: Crear nueva orden
      description: Crea una nueva orden
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Orden creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                type: array
                items:
  /api/orders/{id}:
    put:
      summary: Actualizar orden
      description: Actualiza una orden existente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Orden actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Orden no encontrada
          content:
            application/json:
              schema:
                type: string
    delete:
      summary: Eliminar orden
      description: Elimina una orden existente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Orden eliminada exitosamente
        '404':
          description: Orden no encontrada
    get:
      summary: Comprobar existencia de orden
      description: Verifica si una orden existe por su ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Orden encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Orden no encontrada

components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        productIds:
          type: array
          items:
            type: integer
            format: int64
          example: [1, 2, 3]
        customerName:
          type: string
          example: "Nombre del cliente"
        totalAmount:
          type: number
          format: double
          example: 150.75
        status:
          type: boolean
          example: true